Refactor and deploy Movie client app

+++ src/components/delete-user/deleteUser.jsx
@@ -2,21 +2,28 @@ import { BaseUrl } from "../../constants/constant";
 
 export const DeleteAccount = (event) => {
     event.preventDefault();
-    const token = localStorage.getItem("token");
-    fetch(BaseUrl + "/deleteUser", {
-        headers: {
-            "Authorization": `Bearer ${token}`,
-            "Content-Type": "application/json"
-        },
-        method: "DELETE",
-    }).then((response) => {
-        if (response.ok) {
-            localStorage.clear();
-            window.location.href = "/register";
-        } else {
-            alert("Unable to delete account. Try again");
-        }
-    }).catch(e => {
-        console.log("error: ", e)
-    });
+    const toDelete = window.confirm(
+        "Are you sure you want to delete your account permanently?"
+    )
+
+    if (toDelete) {
+        const token = localStorage.getItem("token");
+        fetch(BaseUrl + "/deleteUser", {
+            headers: {
+                "Authorization": `Bearer ${token}`,
+                "Content-Type": "application/json"
+            },
+            method: "DELETE",
+        }).then((response) => {
+            if (response.ok) {
+                localStorage.clear();
+                window.location.href = "/register";
+            } else {
+                alert("Unable to delete account.");
+            }
+        }).catch(e => {
+            alert("Oops! Something went wrong. Please try again later.");
+        });
+    }
+
 }
